pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "web_watchr"
authors = [{ name = "Tim Adler", email = "tim+github@emrys-merlin.de" }]
description = "Monitors a website for changes and publishes them on telegram"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["bot", "scraping", "telegram"]
license = { text = "MIT" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "python-telegram-bot==21.1",
    "typer==0.12.3",
    "pydantic==2.7.0",
    "loguru==0.7.2",
    "playwright==1.46.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "pre-commit",
    "mypy",
    "ipdb",
]
docs = [
    "mkdocs",
    "mkdocstrings[python]",
]

[tool.setuptools.dynamic]
version = { attr = "web_watchr.__version__" }

# [project.scripts]
# my-script = "my_package.module:function"

[tool.pytest.ini_options]
# minversion = "6.0"
# addopts = "-ra -q"
testpaths = ["test"]
